!play, check if its a possible youtube link. just start playing without arguments

+++ src/main/java/novaz/command/Play.java
@@ -15,8 +15,6 @@ import sx.blah.discord.util.MissingPermissionsException;
 import sx.blah.discord.util.RateLimitException;
 
 import java.io.File;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 /**
  * !play
@@ -24,7 +22,6 @@ import java.util.regex.Pattern;
  * yea.. play is probably not a good name at the moment
  */
 public class Play extends AbstractCommand {
-	private final Pattern yturl = Pattern.compile("^.*((youtu.be/)|(v/)|(/u/\\w/)|(embed/)|(watch\\?))\\\\??v?=?([^#\\\\&\\?]*).*");
 
 	public Play(NovaBot b) {
 		super(b);
@@ -51,40 +48,40 @@ public class Play extends AbstractCommand {
 	public String execute(String[] args, IChannel channel, IUser author) {
 		if (args.length > 0) {
 			boolean justDownloaded = false;
-			String videocode = extractvideocodefromyoutubeurl(args[0]);
-			File filecheck = new File(Config.MUSIC_DIRECTORY + videocode + ".mp3");
-			if (!filecheck.exists()) {
-				IMessage msg = bot.sendMessage(channel, TextHandler.get("music_downloading_hang_on"));
-				YTUtil.downloadfromYoutubeAsMp3(videocode);
-				justDownloaded = true;
-				try {
-					msg.delete();
-				} catch (MissingPermissionsException | RateLimitException | DiscordException e) {
-					e.printStackTrace();
+			String videocode = YTUtil.extractCodeFromUrl(args[0]);
+			if (YTUtil.isValidYoutubeCode(videocode)) {
+				File filecheck = new File(Config.MUSIC_DIRECTORY + videocode + ".mp3");
+				if (!filecheck.exists()) {
+					IMessage msg = bot.sendMessage(channel, TextHandler.get("music_downloading_hang_on"));
+					YTUtil.downloadfromYoutubeAsMp3(videocode);
+					justDownloaded = true;
+					try {
+						msg.delete();
+					} catch (MissingPermissionsException | RateLimitException | DiscordException e) {
+						e.printStackTrace();
+					}
+				}
+				if (filecheck.exists()) {
+					if (justDownloaded) {
+						OMusic rec = TMusic.findByYoutubeId(videocode);
+						rec.title = YTUtil.getTitleFromPage(videocode);
+						rec.youtubecode = videocode;
+						rec.filename = videocode + ".mp3";
+						TMusic.update(rec);
+						bot.addSongToQueue(videocode + ".mp3", channel.getGuild());
+						return ":notes: Found *" + rec.title + "* And added it to the queue";
+					}
+					bot.addSongToQueue(videocode + ".mp3", channel.getGuild());
+					return TextHandler.get("music_added_to_queue");
 				}
 			}
-			if (filecheck.exists()) {
-				if (justDownloaded) {
-					OMusic rec = TMusic.findByYoutubeId(videocode);
-					rec.title = YTUtil.getTitleFromPage(videocode);
-					rec.youtubecode = videocode;
-					rec.filename = videocode + ".mp3";
-					TMusic.update(rec);
-					bot.addSongToQueue(videocode + ".mp3", channel.getGuild());
-					return ":notes: Found *" + rec.title + "* And added it to the queue";
-				}
-				bot.addSongToQueue(videocode + ".mp3", channel.getGuild());
-				return TextHandler.get("music_added_to_queue");
+		} else {
+			if (bot.playRandomSong(channel.getGuild())) {
+				return TextHandler.get("music_started_playing_random");
+			} else {
+				return TextHandler.get("music_failed_to_start");
 			}
 		}
 		return TextHandler.get("music_not_added_to_queue");
-	}
-
-	private String extractvideocodefromyoutubeurl(String url) {
-		Matcher matcher = yturl.matcher(url);
-		if (matcher.find()) {
-			return matcher.group(7);
-		}
-		return url;
 	}
 }
